ajout de multer et mise a jour des controlles

+++ controllers/sauces.js
@@ -1,28 +1,83 @@
 const Sauces = require("../models/sauces");
+const fs = require("fs");
 
 exports.createSauce = (req, res, next) => {
-  delete req.body._id;
-  const sauces = new Sauces({
-    ...req.body,
+  const sauceObject = JSON.parse(req.body.sauce);
+  delete sauceObject._id;
+  delete sauceObject._userId;
+  const sauce = new Sauces({
+    ...sauceObject,
+    userId: req.auth.userId,
+    imageUrl: `${req.protocol}://${req.get("host")}/images/${
+      req.file.filename
+    }`,
   });
-  sauces
+
+  /* likes = 0;
+  dislikes = 0;
+  usersLiked = [];
+  usersDisliked = []*/
+
+  sauce
     .save()
-    .then(() => res.status(201).json({ message: "Sauce enregistrée !" }))
-    .catch((error) => res.status(400).json({ error }));
+    .then(() => {
+      res.status(201).json({ message: "Sauce enregistrée !" });
+    })
+    .catch((error) => {
+      res
+        .status(403)
+        .json({ error, message: "Error 403: unauthorized request" });
+    });
 };
 
 exports.modifySauce = (req, res, next) => {
-  Sauces.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Sauce modifiée !" }))
-    .catch((error) =>
-      res.status(403).json({ error, message: "403: unauthorized request" })
-    );
+  const sauceObject = req.file
+    ? {
+        ...JSON.parse(req.body.sauce),
+        imageUrl: `${req.protocol}://${req.get("host")}/images/${
+          req.file.filename
+        }`,
+      }
+    : { ...req.body };
+
+  delete sauceObject._userId;
+  Sauces.findOne({ _id: req.params.id })
+    .then((sauce) => {
+      if (sauce.userId != req.auth.userId) {
+        res.status(401).json({ message: "Not authorized" });
+      } else {
+        Sauces.updateOne(
+          { _id: req.params.id },
+          { ...sauceObject, _id: req.params.id }
+        )
+          .then(() => res.status(200).json({ message: "Objet modifié!" }))
+          .catch((error) => res.status(401).json({ error }));
+      }
+    })
+    .catch((error) => {
+      res.status(400).json({ error });
+    });
 };
 
 exports.deleteSauce = (req, res, next) => {
-  Sauces.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Sauce supprimé !" }))
-    .catch((error) => res.status(400).json({ error }));
+  Sauces.findOne({ _id: req.params.id })
+    .then((sauce) => {
+      if (sauce.userId != req.auth.userId) {
+        res.status(403).json({ message: "Not authorized" });
+      } else {
+        const filename = sauce.imageUrl.split("/images/")[1];
+        fs.unlink(`images/${filename}`, () => {
+          Sauces.deleteOne({ _id: req.params.id })
+            .then(() => {
+              res.status(200).json({ message: "Objet supprimé !" });
+            })
+            .catch((error) => res.status(401).json({ error }));
+        });
+      }
+    })
+    .catch((error) => {
+      res.status(500).json({ error });
+    });
 };
 
 exports.getAllSauces = (req, res, next) => {
